[add] フォームの前に「新規コインを作成する」ボタンを表示するようにしました。 #24
メニューバーから唐突にフォームに飛ばされるのはどうかと思ったので。
